Rewrite mapN functions to be faster
This makes it so there are not N function calls for mapN. It also means
that all the arguments are applied at once, which can be a lot faster.
Probably not stuff you want to take into account in general, but it
does not hurt to make the standard libraries a bit odd until we have
more advanced optimizations. /cc @laszlopandy