Simpler implementation of Native.List.repeat
The array doubling pattern used by the old implementation probably isn't a good
idea compared to a plain `for` loop that pushes onto an array, because all
(modern) javascript engines dynamically grow arrays in a reasonble way such
that `push` is amortized constant time. There's no need to roll your own
implementation of that, and doing so probably hurts performance.

But in the case of `repeat`, there is no need to create an array at all.