For #221 - The src parameter needs to appear within quotes to work correctly in some cases.
For valid URLs with parentheses in them, for example, the fittedImage and tiledImage
functions won't set the correct style. This quotes the result, assuming that src is a
valid URL encoded string. Javascript's encodeURI function escapes double quotes and
can be used to convert things like file paths to href values.