Revise docs for Mailboxes
Although there are few grammar and code changes, a lot of this is subjective.
Some of it is even formatting style. (I really dislike surround code with
parens; it looks very lisp-y.) Hopefully there are some changes you like, and we
can go back and forth until we work something out.