removed sentence in Dict.elm documentation
The sentence made no sense. The resulting dictionary will only contain the keys that appear in the first but not in the second given dictionary. So of course, the values for those keys can only come from the first dictionary. There is no potential conflict between values, since if a key occurs in both given dictionaries, it will be dropped anyway, so there is no need to decide "by some preference" which value (from first or from second dictionary) to choose.