Fallback to other things when keyCode is not available
Should fix #326, cc @jvoigtlaender

One of the suggestions was to use the same thing as in jQuery:

```javascript
// the jquery one
(event.which != null) ? event.which : (event.charCode != null ?
event.charCode : event.keyCode)
```

But my main concern is that charCode and keyCode and which don’t do
exactly the same thing. At least, it seems that there was a time in the
past that that was true. Furthermore, it feels weird to check if it is
equal to null, because sometimes these things are undefined instead.

So I chose a solution that’ll only fall back to other things if keyCode
is not defined. This means the behavior is definitely the same *except*
when keyCode does not exist. This means we don’t have to hope browsers
are sane to preserve the existing behavior.

I tried to test if `Keyboard.keysDown` and `Keyboard.presses` work
exactly the same with this change. I confirmed this on Chrome, FireFox,
Safari, and Opera on OSX all updated to whatever version is out today.