Change API of Task.perform to not use Never, remove Task.performAndIgnore
I think performAndIgnore is problematic. It’s fine to run a task and
not care about the result, but it is not fine for it to fail silently.
The right API may be more like:

performAndIgnore : (x -> msg) -> Task x () -> Cmd msg

So that users are strongly encouraged to still handle errors.