Refactor Platform implementation for reactor
Need to be able to find out about all the effect managers and ports
that the user has set up. This makes all that information available in
one place and adds a pkg field so we can know if these things are local
or not. For 0.17 the plan is to only allow effect managers that are
defined in dependencies so we can be sure their implementations do not
change.