Improve (==) implementation
Provide custom logic for Dict, Set, and Array for now. This will likely
be handled in a more generic way in a future release of the compiler.

Do way less allocation. The old implementation was allocating an object
for every single pair that needed to be checked for equality. This also
meant using an array push and pop which can be expensive as well. The
new version only pushes work on the stack once we have recursed 100
levels.