Fix #628 - killed processes getting run
Huge thanks to @jagare for pinning down the root cause and providing a
a performant fix. Just need the check if a process is killed before
stepping it. No need to mess with the work queue (which would be more
expensive) or to check on ever loop of step (which is wasted work)