Docs: Note on clamping updates with Time.every
I've run into this while playing with Elm Tutorial and at first though it's a bug in `Time` module. It took me some investigation to find out that it's actually a JS / browser behaviour outside of the Elm. The good thing is that I've learned something new about browsers, but since using `setInterval` in Elm's `Time.every` is an implementation detail, I think this should be explicitly mentioned in the docs.