Move towards simpler scheduler and synchronous app initialization
Stop delaying all tasks by a setTimeout.

Stop having a special case for if there is a specific bug in the
scheduler.

Stop counting numSteps of tasks reductions. The limit was high enough
that I suspect very few folks reached it, and I’m not sure it’s the
right way to approach this in browsers.

It may make sense to revive it with some mechanism for “low priority”
tasks that could be done with requestIdleCallback and timeRemaining

Create the app thread and effect manager processes synchronously. This
avoids the need for the starter queue for incoming ports. It also makes
the code for creating the main thread a bit clearer.