Draft of better API for working with dates and time zones
Result of “Elm Friday” with @stoeffel. Very fun! :D

Future work will involve using
http://howardhinnant.github.io/date_algorithms.html to implement
functions like `year`

We wrote a parser for the basic format used by moment.js, but I think
that was a mistake. I think it’d be best if this module was totally
agnostic to however folks want to populate a `Zone` data structure.