Allow any non-func value in sets or as dict keys
Extends Native.Utils.cmp to also compare records and ADTs; however the
type checker will still require that uses of Basics.compare and
the > < >= <= operators only operate on officially comparable types.

The ordering for ADTs and records is considered arbitrary and internal,
it is implemented just so that they can appear in sets and as
dictionary keys.