Add Random.constant and improve docs a lot
The only way to get an efficient implementation of constant is to put
it in core. All the other stuff can be implemented fine using the
public API.

I also reviewed the random-extra library. It is really nice! I think it
is a bit large to bring other stuff over right now, but I’d like to
circle back after 0.19 and see if it’s possible to end up with a better
reality.

I also redid the docs to give much more helpful examples. The docs also
point people to random-extra quite a lot, so they will be more likely
to find the helper functions they need.