replace refernce to `%` with `modBy`
I just ran into a runtime error caused by calling `Basics.modBy` with the arguments in the wrong order, leading to a value of 0 as the first argument. The error had a link to [the 1.0.0 version of the page i edited](https://github.com/elm/core/blob/1.0.0/hints/11.md). My understanding is that `%` has been removed from the language, in favor of `remainderBy` and `modBy`, so, given that the old version of the page would still be around for old versions of elm, changing this to reference `modBy` seems better. 

As mentioned in [this issue comment](https://github.com/elm/core/issues/1034#issuecomment-503159937), `remainderBy 0 n` currently returns `NaN` instead of leading to this page, so I didn't mention it.